+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 * Update Node.js to v18 in GitHub Actions. Refs UIIN-2520.
 * Fix issue with Relevance sort option. Refs UIIN-2474.
 * Change tenant id to central when opening details of shadow instance. Refs UIIN-2529.
+* ECS: use mod-search to get tenantId for local/shared instance details. Refs UIIN-2538.
 
 ## [9.4.11](https://github.com/folio-org/ui-inventory/tree/v9.4.11) (2023-08-02)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.10...v9.4.11)